Rename posts/Feed* -> posts/PostFeed*
